renamed simulate pipeline
