+++ sb_pipe/pipelines/simulate/simulate__plot_sim_exp_error_bars.r
@@ -27,7 +27,7 @@
 
 # Retrieve the environment variable SB_PIPE
 SB_PIPE <- Sys.getenv(c("SB_PIPE"))
-source(file.path(SB_PIPE, 'sb_pipe','pipelines','sb_simulate','plot_sim_exp_func.r'))
+source(file.path(SB_PIPE, 'sb_pipe','pipelines','simulate','plot_sim_exp_func.r'))
